feat: make custom hook of details page and all logic define in hook
