+++ src/app/(routes)/employee/details/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Sidebar from "../../../components/sidebar/Sidebar";
 import Header from "../../../header/Header";
 import Table from "../../../components/table/Table";
@@ -7,103 +7,25 @@ import Pagination from "../../../components/pagination/Pagination";
 import RowActions from "../../../components/rowActions/RowActions";
 import Image from "next/image";
 import EmployeeTableToolbar from "../../../components/employeeTableToolbar/EmployeeTableToolbar";
-import { toast } from "react-toastify";
-import { useRouter } from "next/navigation";
 import { Employee } from "@/app/types/types";
+import { useEmployees } from "./useEmployees";
 
 function Page() {
-  const router = useRouter();
-  const [employees, setEmployees] = useState<Employee[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [recordsPerPage, setRecordsPerPage] = useState(5);
-  const [searchQuery, setSearchQuery] = useState("");
-  const [totalRecords, setTotalRecords] = useState(0);
-
-  const fetchEmployees = async () => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const response = await fetch("/api/employee");
-      if (!response.ok) {
-        throw new Error("Failed to fetch employees");
-      }
-
-      const data = await response.json();
-      setEmployees(data);
-      setTotalRecords(data.length);
-      setIsLoading(false);
-    } catch (err: any) {
-      setError(err.message || "An error occurred while fetching employees");
-      toast.error("Failed to load employee data");
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchEmployees();
-  }, []);
-
-  const handleSearch = (query: string) => {
-    setSearchQuery(query);
-    setCurrentPage(1);
-  };
-
-  const handleFilter = () => {
-    toast.info("Filter functionality will be implemented here");
-  };
-
-  const handleEdit = (employee: Employee) => {
-    router.push(`/admin/employees/edit/${employee.id}`);
-  };
-
-  const handleDelete = async (employee: Employee) => {
-    if (
-      confirm(
-        `Are you sure you want to delete ${employee.firstName} ${employee.lastName}?`
-      )
-    ) {
-      try {
-        const response = await fetch(`/api/employee?id=${employee.id}`, {
-          method: "DELETE",
-        });
-
-        if (!response.ok) {
-          throw new Error("Failed to delete employee");
-        }
-
-        toast.success("Employee deleted successfully");
-        fetchEmployees();
-      } catch (err: any) {
-        toast.error(err.message || "Failed to delete employee");
-      }
-    }
-  };
-
-  const handleView = (employee: Employee) => {
-    router.push(`/viewemployeedetail${employee.id}`);
-  };
-
-  const filteredEmployees = employees.filter((employee) => {
-    if (!searchQuery) return true;
-
-    const fullName = `${employee.firstName} ${employee.lastName}`.toLowerCase();
-    const searchLower = searchQuery.toLowerCase();
-
-    return (
-      fullName.includes(searchLower) ||
-      employee.employeeId.toLowerCase().includes(searchLower) ||
-      employee.workEmail.toLowerCase().includes(searchLower) ||
-      employee.department.toLowerCase().includes(searchLower)
-    );
-  });
-
-  const startIndex = (currentPage - 1) * recordsPerPage;
-  const paginatedData = filteredEmployees.slice(
-    startIndex,
-    startIndex + recordsPerPage
-  );
+  const {
+    isLoading,
+    error,
+    currentPage,
+    recordsPerPage,
+    filteredEmployees,
+    paginatedData,
+    handleSearch,
+    handleFilter,
+    handleEdit,
+    handleDelete,
+    handleView,
+    setCurrentPage,
+    setRecordsPerPage,
+  } = useEmployees();
 
   const columns = [
     {
@@ -164,12 +86,11 @@ function Page() {
           <Sidebar />
         </div>
         <div className="w-full">
-          <div className="w-full">
-            <Header
-              title="All Employees"
-              description="All Employee Information"
-            />
-          </div>
+          <Header
+            title="All Employees"
+            description="All Employee Information"
+          />
+
           <div className="max-h-[86vh] w-full bg-transparent border border-[#A2A1A833] rounded-[10px] p-4 flex flex-col">
             <div className="mb-4 flex-shrink-0">
               <EmployeeTableToolbar
@@ -199,7 +120,7 @@ function Page() {
                 currentPage={currentPage}
                 recordsPerPage={recordsPerPage}
                 totalRecords={filteredEmployees.length}
-                onPageChange={(page) => setCurrentPage(page)}
+                onPageChange={setCurrentPage}
                 onRecordsPerPageChange={(newRecords) => {
                   setRecordsPerPage(newRecords);
                   setCurrentPage(1);
